requirements.txt
Django>=3.0,<4.0
djangorestframework==3.12.2
psycopg2-binary>=2.8
# psycopg2==2.7.3.2
drf-yasg==1.20.0
Pillow==2.2.1
djangorestframework-simplejwt